Move to object.cpp/object.h
